[front/components/members] - refactor: centralize member role handling logic
- Moved logic for member role changes from `InvitationModal.tsx` and `members/index.tsx` to `lib/client/members.ts`
- Updated `handleMembersRoleChange` usage to match the new centralized function signature

[front/lib/api] - fix: ensure updateInvitationStatusAndRole requires role

- Made `role` parameter mandatory when updating invitation status and role to prevent potential inconsistencies

[front/pages/api] - refactor: adjust API handler to comply with updated invitation update logic

- Aligned the API handler `PostMemberInvitationBodySchema` to enforce a non-optional `initialRole` field for consistency
